Talleres/tareaLLM/requirements.txt
langchain
langchain-community
openai
python-dotenv
tiktoken
langchain_openai
pinecone
langchain-text-splitters
langchain-huggingface 
sentence-transformers
pinecone-client